me falla lo de components ns porque

+++ RazorPages/RazorPages/Pages/Alumnos/Edit.cshtml.cs
@@ -1,3 +1,4 @@
+using System.Diagnostics.Eventing.Reader;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
@@ -9,10 +10,12 @@ namespace RazorPages.Pages.Alumnos
     public class EditModel : PageModel
     {
         private readonly IAlumnoRepositorio AlumnoRepositorio;
-		private readonly IWebHostEnvironment webHostEnvironment;
+        private readonly IWebHostEnvironment webHostEnvironment;
 
-		public Alumno alumno { get; set; }
+        [BindProperty]
+        public Alumno alumno { get; set; }
 
+        [BindProperty]
         public IFormFile Photo { get; set; }
 
         public EditModel(IAlumnoRepositorio alumnoRepositorio, IWebHostEnvironment webHostEnvironment)
@@ -20,27 +23,54 @@ namespace RazorPages.Pages.Alumnos
             this.AlumnoRepositorio = alumnoRepositorio;
             this.webHostEnvironment = webHostEnvironment;
         }
-        public void OnGet(int id)
+        public IActionResult OnGet(int? id)
         {
-            alumno = AlumnoRepositorio.GetAlumno(id);
+            if (id.HasValue)
+            {
+                alumno = AlumnoRepositorio.GetAlumno(id.Value);
+
+            }
+            else
+            {
+                alumno = new Alumno();
+            }
+            return Page();
         }
 
         public IActionResult OnPost(Alumno alumno)
         {
-            if(Photo !=  null)
+            if (ModelState.IsValid)
             {
-                if(alumno.Foto != null)
-                {
-					string filePath = Path.Combine(webHostEnvironment.WebRootPath, "images", alumno.Foto);
-					System.IO.File.Delete(filePath);
-				}
-                alumno.Foto = ProcessUploadedFile();
+                if (Photo != null) { 
+                    if (alumno.Foto != null)
+                    {
+                        string filePath = Path.Combine(webHostEnvironment.WebRootPath, "images", alumno.Foto);
+                        System.IO.File.Delete(filePath);
+                    }
+                    alumno.Foto = ProcessUploadedFile();
+                }   
+
+            if (alumno.Id > 0)
+            {
+                AlumnoRepositorio.Update(alumno);
             }
-
-
-            AlumnoRepositorio.Update(alumno);
+            else
+            {
+                AlumnoRepositorio.Add(alumno);
+            }
             return RedirectToPage("Index");
         }
+            else
+            {
+                return Page();
+            }
+
+        }
+           
+            
+
+        
+    
 		private string ProcessUploadedFile()
 		{
 			//necesitamos un objeto de una clase que sea capaz de manipular el proyecto por lo que la creamos en el constructor
